+++ ios/linkid_mmp.xcframework/ios-arm64/linkid_mmp.framework/Headers/linkid_mmp-Swift.h
@@ -261,6 +261,7 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 #pragma clang diagnostic ignored "-Watimport-in-framework-header"
 #endif
 @import CoreFoundation;
+@import Foundation;
 @import UIKit;
 #endif
 
@@ -282,6 +283,15 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 #endif
 
 #if defined(__OBJC__)
+@class UIApplication;
+@class NSURL;
+
+SWIFT_CLASS("_TtC10linkid_mmp23AirflexDeepLinkDelegate")
+@interface AirflexDeepLinkDelegate : UIResponder <UIApplicationDelegate>
+- (BOOL)application:(UIApplication * _Nonnull)application openURL:(NSURL * _Nonnull)url options:(NSDictionary<UIApplicationOpenURLOptionsKey, id> * _Nonnull)options SWIFT_WARN_UNUSED_RESULT;
+- (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
+@end
+
 @class UIEvent;
 @class UIWindowScene;
 @class NSCoder;